Migrate String and Markdown tests to test_utils
